🍣seperate test directory #28
