+++ README.md
@@ -56,3 +56,5 @@ lets enjoy the polls by me :)
 * [Iteration 3 plan](https://github.com/Tezigudo/ku-polls/wiki/Iteration-3-Plan)
 
 [django-tutorial]: https://docs.djangoproject.com/en/4.1/intro/tutorial01/sdx
+<!-- user: banana
+pass: iloveapple -->
