+++ polls/models.py
@@ -4,6 +4,8 @@ import datetime
 from django.contrib import admin
 from django.db import models
 from django.utils import timezone
+from django.contrib.auth.models import User
+
 
 
 class Question(models.Model):
