+++ polls/tests/test_detailview.py
@@ -1,0 +1,30 @@
+from .base import create_question
+
+from django.test import TestCase
+
+
+
+class QuestionDetailViewTests(TestCase):
+    def test_future_question(self):
+        """
+        The detail view of a question with a pub_date in the future
+        returns a 302 redirected (unpublished question)
+        """
+        future_question = create_question(
+            question_text='Future question.', days=5)
+        url = reverse('polls:detail', args=(future_question.id,))
+        response = self.client.get(url)
+        # if polls arent open it will reqirect(maybe) or
+        # 302 redirect i dont sure this one
+        self.assertEqual(response.status_code, 302)
+
+    def test_past_question(self):
+        """
+        The detail view of a question with a pub_date in the past
+        displays the question's text.
+        """
+        past_question = create_question(
+            question_text='Past Question.', days=-5)
+        url = reverse('polls:detail', args=(past_question.id,))
+        response = self.client.get(url)
+        self.assertContains(response, past_question.question_text)
