+++ polls/tests/test_voting.py
@@ -1,0 +1,4 @@
+from .base import create_question
+
+from django.test import TestCase
+
